refactor: 플랜 vo 패키지 생성 및 이동 (#135)
