+++ src/main/kotlin/kr/wooco/woocobe/plan/domain/vo/PlanPlace.kt
@@ -1,4 +1,4 @@
-package kr.wooco.woocobe.plan.domain.entity
+package kr.wooco.woocobe.plan.domain.vo
 
 data class PlanPlace(
     val order: Int,
