+++ src/main/kotlin/kr/wooco/woocobe/plan/domain/vo/PlanRegion.kt
@@ -1,0 +1,6 @@
+package kr.wooco.woocobe.plan.domain.vo
+
+data class PlanRegion(
+    val primaryRegion: String,
+    val secondaryRegion: String,
+)
